Plugins: port xunit plugin
